cuda image processing

+++ src/imgui/imgui_saiga.cpp
@@ -3,7 +3,7 @@
 namespace ImGui {
 
 void Texture(std::shared_ptr<Saiga::raw_Texture> texture, const ImVec2& size){
-    ImGui::Image((void*)texture->getId(),size,vec2(0,1),vec2(1,0));
+    ImGui::Image((void*)(intptr_t)texture->getId(),size,vec2(0,1),vec2(1,0));
 }
 
 }
